TP3/Hit1/requirements.txt
pika
fastapi
uvicorn
Pillow
